refactoring, extract getMergeStatus() method to a new class
